repairChecks uses tx. delete record test case. sqlboiler update.

+++ model/insert.go
@@ -154,7 +154,7 @@ func upload(f models.File, values url.Values, key string) (models.File, error) {
 	f.DateIssuedMonth = month
 	f.DateIssuedDay = day
 	f.Filename = filename
-	f.Filesize = int64(filesize)
+	f.Filesize = filesize
 	f.FileZipContent = content
 	f.RetrotxtReadme = readme
 	f.FileMagicType = filemagic
